Fixes race condition on select input validation
When a select type input had it's value changed, the property component
was calling the valid function before the properties had been updated
resulting in the validation message showing, even though the state of
the component was valid. This fix moves the valid check into the
componentDidUpdate lifecycle callback. It checks if the value has
changed, and if so it calls the validation method.