Fixes undefined on change function in form component
This fixes the undefined function error that was happening in the form
component. In a previous commit, I passed the onchange function on
through to the form component, but didn't define the onchange on the
form component. This fixes that.

It removes the default onchange prop, and checks for the existence of
an onchange prop, and passes the serialized form data through if it
exists, otherwise it does nothing.